services/auth-service/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
python-dotenv==1.0.0
sqlalchemy==2.0.12
alembic==1.10.4
psycopg2-binary==2.9.6
redis==4.5.5
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
httpx==0.24.0
tenacity==8.2.2
pytest==7.3.1
starlette==0.27.0
prometheus-client==0.16.0
email-validator==2.0.0 